- Added dataswarm_blob to represent data structure. - Fixed up dataswarm_blob_table to store metadata in filesystem. - Eliminated the rw/ro directories to reflect blob state.
